redis-cluster/node-app/requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
redis==5.0.1

